requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
streamlit
python-dotenv
requests
passlib[bcrypt]
pytest
 httpx
